--- packages/cypress-cloud/lib/bus.ts
@@ -1,2 +1,0 @@
-import { EventEmitter } from "events";
-export const bus = new EventEmitter();
